issueStatement.java - added a method that passes back to the calling class a result set as a hash table where the attribute name is a key and the returned value is the value.
queryStore.java - added a new method, that will replace the getPlotSummary
method, that returns the summary of a plot using as input a plot ID number
and by querying the de-normalixed summary tables that will be updated everynight.

DbConnectionBroker.java - nothing changed

sqlMapper.java -- now uses the new getPlotSummary method in the queryStore
class

xmlWriter.java -- modified this class to use the PlotDataMapper and
PlotXmlWriter classes - this class will be removed or renamed soon.


CVS : ----------------------------------------------------------------------