Re-checkin of this class file -- the CVS sub-dir was lost during the argus server crash.  I fixed the numeric data type in the class so that it now is correctly updated in the sql files generated by the torque program.  This fix included editing this class and updating the data-typ map in:
templates/sql/base/postgres/db.props

below are the former logs:
----------------------------
revision 1.6
date: 2001/10/20 00:52:25;  author: harris;  state: Exp;  lines: +7 -3
Added the new tables in the plant/community taxonomy components of the database
that Bob has added over the last couple days to generate the correct sql
----------------------------
revision 1.5
date: 2001/10/17 20:40:36;  author: harris;  state: Exp;  lines: +60 -8
Added functionality to handle the plant-taxonomy component of the data model, the
class is pretty much done at this point evertything is working as is.  The next
step will be to genetrate all the tables (including the plots, plant-tax, and comm-tax)
in the same space.
----------------------------
revision 1.4
date: 2001/10/17 18:34:30;  author: harris;  state: Exp;  lines: +100 -9
Added methods to handle the community taxonomy component of the data model.
----------------------------
revision 1.3
date: 2001/10/17 16:48:27;  author: harris;  state: Exp;  lines: +147 -16
Added the facility to automatically create torque-xml document given as input the directory that
contains the vegclass-xml documentation.
----------------------------
revision 1.2
date: 2001/10/16 18:33:17;  author: harris;  state: Exp;  lines: +27 -8
Added a second constructor which will hendle all the tables in a database
and atomized the constructor by creating more methods
----------------------------
revision 1.1
date: 2001/10/16 04:15:00;  author: harris;  state: Exp;
This class hadles the transformation of an xml document consistent with the
format that we have been using for documenting the database into other formats.
At this point it does the transformation to the format that is consistent with the
jakarta-torque dtd.  The torque tool will take an xml document that defines a database,
or component of a database and produce database specific code.  I'm interested in
using this tool (torque) because it supports oracle, postgresql and hypersonic sql -
the three databases that I am using for my project.
=============================================================================