This package is experimental!  Basically it allows a user to access plot data sources (including vegbank, tnc, ncvs, turboveg, etc..) accross a network using the RMI api.   There are a number of reasons that I think that this package will become important:
1] many of the plots that I have worked with are stored in ms access formats, which
eventhough this means the user has to have access to access they are easire to parse
than when they are in an ascii table b/c JDBC can be used to direct sql statements directly
to the source.  Furthermore B. Peet is really keen on building an access template for
integrating and storing plots data, which is fine B/C he is going to get someone else to do it
and all I will have to do is to write the java-plugin to access the data.  So since there is so
reliance on access and I will not use any MS OS I need a tool to access these data sources
on the MS platform.  In the future the linux odbc source stuff should be looked into.

2]  Bob has decided to really emphasize the the vegbank server, and anticipates that
it will provide users the complete functionality that they need to manage their plots.
This means that the emhasis on an embedded system is less, which is cool because that
is a huge amount of work that will no way be done by July.  Using RMI and either an
applet or an RMI Client Application the user will be provided with a seriously interactive
system for working with their plots directly on the server.  This could be done using http/xml
but I thought that this RMI biz is easier to work with.  In the future one should look at using
SOAP or a web-dav type protocol(s).

As mentioned above this is pretty experimental, but I ma planning on having something to
show for the working group meeting next week that will allow a user to login to vegbank,
upload their data, use an applet (or possibly the client) to connect massage their data
and then load their data (where the massaging and the loading will be done with RMI calls)