development.html no longer needed, because replaced by:
new xml-xsl approach for development page, allows table to be alternately coloured; more easily maintainable.  Edit XML as you would normal HTML (but keep it balanced!) for development.xml.  Just leave the bottom table as "tableToStyle" as this is what the xsl uses to know that this table should be styled by it.