First attempt to get this build working in with the cvs directory moves.
Dropped calls to veg_plot, veg_community and veg_taxa builds.

Take the context name as a user input, this could get irksome in development so I suggest this workaround:
% echecho "[build.context]" | ant [target]o "[build.context]" | ant [target]

compile, jar and deploy appear to work for both vegbank and vegtest.

test, report and dist tasks not working/checked.